\name{FuzzyToolkitUoN-package}
\alias{FuzzyToolkitUoN-package}
\alias{FuzzyToolkitUoN}
\docType{package}
\title{
Type 1 Fuzzy Logic Toolkit
}
\description{
A custom framework for working with Type 1 Fuzzy Logic, produced by the University of Nottingham IMA Group.
}
\details{
\tabular{ll}{
Package: \tab FuzzyToolkitUoN\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-04-26\cr
License: \tab GPL (>= 2)\cr
NeedsCompilation: \tab No\cr
Repository: \tab CRAN\cr
}
}

\author{
Craig Knott, Luke Hovell, Nathan Karimian

Maintainer: Jon Garibaldi <jon.garibaldi@ima.ac.uk>
}